Use download_dependency to download node
- remove all usages of tranlate_dependency_url

[#128099189]

+++ lib/binaries.sh
@@ -24,10 +24,20 @@ install_nodejs() {
   else
     echo "Downloading and installing node $resolved_version..."
   fi
+
   local heroku_url="https://s3pository.heroku.com/node/v$resolved_version/node-v$resolved_version-$os-$cpu.tar.gz"
-  local download_url=`translate_dependency_url $heroku_url`
-  local filtered_url=`filter_dependency_url $download_url`
-  curl "$download_url" --silent --fail --retry 5 --retry-max-time 15 -o /tmp/node.tar.gz || (>&2 $BP_DIR/compile-extensions/bin/recommend_dependency $heroku_url && false)
+  local exit_code=0
+  local filtered_url=""
+
+  filtered_url=$($BP_DIR/compile-extensions/bin/download_dependency $heroku_url /tmp) || exit_code=$?
+  if [ $exit_code -ne 0 ]; then
+    echo -e "`$BP_DIR/compile-extensions/bin/recommend_dependency $heroku_url`" 1>&2
+    exit 22
+  fi
+
+  local downloaded_file=$(ls /tmp/node-v*.tar.gz)
+  mv $downloaded_file /tmp/node.tar.gz
+
   echo "Downloaded [$filtered_url]"
   tar xzf /tmp/node.tar.gz -C /tmp
   rm -rf $dir/*
